parser/requirements.txt
aiohttp>=3.9.0
aiolimiter>=1.1.0
aiofiles>=23.0.0
Pillow>=10.0.0
imagehash>=4.3.0
certifi>=2023.0.0
